Make interop platform-specific
